app/requirements.txt
# User Microservice Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

